# perf: perform blobless clones in git enumeration